+++ CHANGELOG.md
@@ -5,6 +5,7 @@ file.
 
 ## [Unreleased]
 ### Added
+- Added Benchmarks to `RunType` to allow coverage of benchmark tests to be collected
 
 ### Changed